+++ cashu-lib-common/src/main/java/xyz/tcheeric/cashu/common/Proof.java
@@ -4,12 +4,14 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import xyz.tcheeric.cashu.common.json.deserializer.SecretDeserializer;
 import xyz.tcheeric.cashu.common.json.deserializer.WitnessDeserializer;
+import xyz.tcheeric.cashu.common.json.serializer.SecretSerializer;
 import xyz.tcheeric.cashu.common.nut12.DLEQProof;
 
 @Data
@@ -24,6 +26,7 @@ public class Proof<T extends Secret> {
 
     @JsonProperty
     @JsonDeserialize(using = SecretDeserializer.class)
+    @JsonSerialize(using = SecretSerializer.class)
     private T secret;
 
     @JsonProperty("id")